backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-dotenv==1.1.0
google-genai
aiofiles==25.1.0
PyJWT[crypto]
